fixed bug #932322
Now, when clicking on "Native editing" with an entry that has not been inserted in the
db so far, an entry is created and inserted, so that the native editor does not start
empty.

This has the side effect of having a new entry even if the editing is later cancelled.

imported from: gobry@pybliographer.org--public/pybliographer--stable--1.2--patch-100